src/players: fix getting cards.
The cards played was not always the good ones. Indeed, on trump round
(that's say, the asked color is the trump color), players has to play a
better card if they can.
Moreover, it hr player does not have the asked color (that is not the
trump color) but have a card of the trump color, he has to cut if his
teammate is not the current master of the trick. If he cut and another
player has also cut, the player has to play a better card if he can.

These cases was not handled. They are now.