src/players: improve 'remove_elem_card_from_player'.
Improve the signature of the method 'remove_elem_card_from_player'.
On this method was given, the player, the element to remove and the
color of the card associated to the element to remove.
But the card, and so the color, can be directly got fom the element
given in argument.
So, it is just simpler and better to just give the element to remove.