src/players: fix memory leak.
When getting a card from a player, we removed the associated element
from the list of the associated color of the player. But, the element
containing the pointer of this card should not just be removed, but also
deleted.
To be perfectly clear, the card will not be deleted, but only the list
element.